requirements.txt
django
djangorestframework
django-taggit
coreapi==2.3.3  # https://github.com/core-api/python-client
django-cors-headers